+++ tests/testthat/test-02-misc.R
@@ -1,0 +1,8 @@
+context('misc functions')
+
+test_that('compatibility with coda package', {
+	library(coda)
+	x <- mcmc.list(mcmc(rnorm(100)),mcmc(rnorm(100)))
+	g <- gelman.diag(x)
+	print(g)
+})