+++ NEWS.md
@@ -1,3 +1,7 @@
+# bayescanr 1.0.1
+* fixed bug where files for multiple replicates would overwrite each other
+* added funcionality to produce traceplots and calculate Gelman-Rubin statistics using coda
+
 # bayescanr 1.0.0
 
 * Basic functionality